feat(edupluss/api) + featfeat(edupluss/educlient) userProgress created, ratings created, google forms link created

+++ api/src/controllers/Steps/deleteStep.js
@@ -1,17 +1,33 @@
-const { Step } = require('../../db');
-const { catchedAsync } = require('../../utils');
+const { Activity, Step } = require("../../db");
+const { catchedAsync } = require("../../utils");
 
 const deleteStep = async (req, res) => {
   const { id } = req.params;
   try {
-    const deleteStep = await Step.destroy({
+    const step = await Step.findByPk(id);
+    if (!step) {
+      return res.status(404).json({ error: "Step not found" });
+    }
+
+    const activity = await Activity.findByPk(step.activityId);
+    if (!activity) {
+      return res.status(404).json({ error: "Activity not found" });
+    }
+
+    if (activity.numberSteps > 0) {
+      activity.numberSteps -= 1;
+      await activity.save();
+    }
+
+    await Step.destroy({
       where: {
         id,
       },
     });
-    res.status(200).json(deleteStep);
+
+    res.status(200).json({ message: "Step deleted successfully" });
   } catch (error) {
-    res.status(404).send(error.message);
+    res.status(500).json({ error: "Internal server error" });
   }
 };
